requirements.txt
Django==1.7.6
Markdown==2.6.1
argparse==1.2.1
django-filter==0.9.2
djangorestframework==3.1.0
wsgiref==0.1.2
